Add files via upload
Folder for the test version of the site. Download this and edit the code on your own computer, and make sure to either message/email everyone else working on the website as well as Bill and Guy with the process you took as well as the edits you made. Thank you.