Update index.css
Updated the navbar with proper spacing, logo colors, and dropdown menus. 

The only issue is that the dropdown menu for the "shop" button is being difficult, as in the animation for hovering over the dropdown menu's buttons is very jittery. (Maybe has something to do with the width and/or positioning of itself and/or other elements? That's my best guess.)