Overhauled Home Page
Complete overhaul of the home page. Still a pretty barebones WIP, but the nav bar is functional and is separated from the main CSS page, the spacing between and within the elements now utilizes the font size (em), and finally, the <main> and <footer> elements are now implemented and ready.