ACTUALLY updated home page
Yeah, I don't know how to use GitHub. I accidentally uploaded my updated version of the home page as three separate HTML and CSS files into the main directory.

Anyways, I made a complete overhaul of the home page! It's still a pretty barebones WIP, but the nav bar is functional and is separated from the main CSS page, the scale of certain elements and the spacing between and within said elements now utilizes the font size (em), and finally, the <main> and <footer> elements are now implemented and ready.